Properly display build errors on ST3 <r3153
As of r3153, Sublime Text's built-in ExecCommand sends strings to on_data, and
as a string must be sent to on_finished with the formatted build output. If
we're on an older version of Sublime Text, therefore, we need to keep track of
whether we have received strings or bytes and send the right data format to
on_finished.