Initial support for the Elm 0.19 compiler
Drops support for Elm 0.18 for now, since it is so different. Also drops support for autocompletion and displaying type information with elm-oracle, which does not yet support Elm 0.19.

For now, supporting only a single build type (building the current file to /dev/null). The full project build will return in a later version.